fix: sort by dollar
* fix: replace solvBTC image

* fix: order asset list for user balance in dollar value

* fix: create analytics components, fix for amplitude
