+++ frontend/app/components/molecules/modals/ModalSelectAsset.tsx
@@ -44,9 +44,16 @@ const ModalSelectAsset: React.FC<ModalSelectAssetProps> = ({ onSelectAsset, onCl
             token.denom.toLowerCase().includes(e.target.value.toLowerCase())
           );
         })
-        .sort(
-          (a, b) => getNumericBalance(b.denom, b.decimals) - getNumericBalance(a.denom, a.decimals),
-        ),
+        .sort((a, b) => {
+          return (
+            getPrice(getNumericBalance(b.denom, b.decimals), b.denom, {
+              format: false,
+            }) -
+            getPrice(getNumericBalance(a.denom, a.decimals), a.denom, {
+              format: false,
+            })
+          );
+        }),
     );
   };
 
@@ -64,36 +71,47 @@ const ModalSelectAsset: React.FC<ModalSelectAssetProps> = ({ onSelectAsset, onCl
           />
         </div>
         <div className="flex flex-col gap-2">
-          {filteredTokens.map((token, index) => {
-            return (
-              <div
-                key={index}
-                className="flex items-center gap-3 rounded-2xl px-3 py-3 justify-between hover:bg-white/10 transition-all cursor-pointer"
-                onClick={() => [hideModal(), onSelectAsset(token)]}
-              >
-                <div className="flex gap-3 items-center">
-                  <img src={token.logoURI} alt={token.symbol} className="h-8 w-8" />
-                  <div className="flex flex-col min-w-0 gap-1">
-                    <p>{token.symbol}</p>
-                    <TruncateText text={token.denom} className="text-sm text-white/50" />
+          {filteredTokens
+            .sort((a, b) => {
+              return (
+                getPrice(getNumericBalance(b.denom, b.decimals), b.denom, {
+                  format: false,
+                }) -
+                getPrice(getNumericBalance(a.denom, a.decimals), a.denom, {
+                  format: false,
+                })
+              );
+            })
+            .map((token, index) => {
+              return (
+                <div
+                  key={index}
+                  className="flex items-center gap-3 rounded-2xl px-3 py-3 justify-between hover:bg-white/10 transition-all cursor-pointer"
+                  onClick={() => [hideModal(), onSelectAsset(token)]}
+                >
+                  <div className="flex gap-3 items-center">
+                    <img src={token.logoURI} alt={token.symbol} className="h-8 w-8" />
+                    <div className="flex flex-col min-w-0 gap-1">
+                      <p>{token.symbol}</p>
+                      <TruncateText text={token.denom} className="text-sm text-white/50" />
+                    </div>
+                  </div>
+                  <div className="flex flex-col gap-1 text-end">
+                    <p>
+                      {convertMicroDenomToDenom(
+                        getBalance(token.denom),
+                        token.decimals,
+                        token.decimals === 6 ? 2 : token.decimals,
+                        false,
+                      )}
+                    </p>
+                    <p className="text-sm text-white/50">
+                      {getPrice(getNumericBalance(token.denom, token.decimals), token.denom)}
+                    </p>
                   </div>
                 </div>
-                <div className="flex flex-col gap-1 text-end">
-                  <p>
-                    {convertMicroDenomToDenom(
-                      getBalance(token.denom),
-                      token.decimals,
-                      token.decimals === 6 ? 2 : token.decimals,
-                      false,
-                    )}
-                  </p>
-                  <p className="text-sm text-white/50">
-                    {getPrice(getNumericBalance(token.denom, token.decimals), token.denom)}
-                  </p>
-                </div>
-              </div>
-            );
-          })}
+              );
+            })}
         </div>
       </div>
     </BasicModal>
